refactor: 검증 관련 클래스 패키지 이동 및 어노테이션 변경
