+++ src/main/java/gamara/server/validator/BasicValidator.java
@@ -1,15 +1,13 @@
-package gamara.server.service;
+package gamara.server.validator;
 
 import gamara.server.common.exception.AppException;
 import gamara.server.common.exception.ErrorCode;
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Service;
+import org.springframework.stereotype.Component;
 
-@Service
+@Component
 @RequiredArgsConstructor
-@Slf4j
-public class ValidateService {
+public class BasicValidator {
 
     public void checkValidIdRange(long id) {
         if (id <= 0) {
